# docs/spec/api/openapi.yaml
openapi: 3.1.0
info:
  title: IEapp API
  version: "0.1"
  description: |
    Minimal OpenAPI document for the current REST surface.
servers:
  - url: http://localhost:8000
paths:
...
  /api/workspaces/{workspace_id}:
    get:
      summary: Get workspace
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    patch:
      summary: Patch workspace
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /api/workspaces/{workspace_id}/test-connection:
    post:
      summary: Test storage connection
      parameters:
        - name: workspace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...